+++ spec/resources/dnsmasq_local_app.rb
@@ -6,13 +6,24 @@ shared_context 'resources::dnsmasq_local_app' do
   include_context 'resources'
 
   let(:resource) { 'dnsmasq_local_app' }
-  let(:properties) { {} }
+  %i(version).each { |p| let(p) { nil } }
+  let(:properties) { { version: version } }
   let(:name) { 'default' }
 
   shared_examples_for 'any platform' do
     context 'the default action (:install)' do
-      it 'installs the Dnsmasq package' do
-        expect(chef_run).to install_package('dnsmasq')
+      context 'the default attributes' do
+        it 'installs the Dnsmasq package' do
+          expect(chef_run).to install_package('dnsmasq')
+        end
+      end
+
+      context 'an overridden version attribute' do
+        let(:version) { '1.2.3' }
+
+        it 'installs the requested version of the Dnsmasq package' do
+          expect(chef_run).to install_package('dnsmasq').with(version: '1.2.3')
+        end
       end
     end